backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
mongoengine>=0.27.0
django-cors-headers>=4.0
celery>=5.3
redis>=5.0
channels>=4.0
channels-redis>=4.0
daphne>=4.0
PyJWT>=2.8
requests>=2.31
python-dotenv>=1.0
pymongo>=3.12.3
